Implementing avro generator to send messages to the queue

+++ ApacheKafkaBasics/Interfaces/IShoppingCart.cs
@@ -1,4 +1,4 @@
-using ApacheKafkaBasics.Models;
+using Generated.Entity;
 
 namespace ApacheKafkaBasics.Interfaces;
 
@@ -6,6 +6,6 @@ public interface IShoppingCart
 {
     public Task<bool> AddProduct(Product product, int quantity);
     public Task<bool> RemoveProduct(int productId);
-    public Task<decimal> GetTotalPrice();
+    public Task<double> GetTotalPrice();
     public Task<List<CartItem>> GetCartItems();
 }
